Changed the project's file structure
