+++ .gitignore
@@ -1,3 +1,6 @@
 .idea/
 .vagrant/
+env/
+venv/
+*/__pycache__/*
 nohup.out
