+++ baboon/static/favicon.ico
